Сделал один универсальный генератор отчетов.
