+++ src/excel_generator/builder.py
@@ -15,6 +15,7 @@ class AnalyticReportBuilder:
 
     async def generate_report(
         self,
+        description: str,
         data: tuple[TasksAnalyticReportDto],
         workbook: Workbook,
         analytic_task_report_full: BaseAnalyticReportSettings,
@@ -22,6 +23,7 @@ class AnalyticReportBuilder:
         """Генерация листа с данными."""
         worksheet = self._create_sheet(workbook, sheet_name=analytic_task_report_full.sheet_name)
         analytic_task_report_full.row_count = 0
+        self.__add_description(worksheet, description, analytic_task_report_full)
         self.__add_header(worksheet, analytic_task_report_full)
         self.__add_data(worksheet, data, analytic_task_report_full)
         self.__add_footer(worksheet, analytic_task_report_full)
@@ -58,6 +60,15 @@ class AnalyticReportBuilder:
         """Создаёт лист внутри отчёта."""
         return workbook.create_sheet(sheet_name)
 
+    def __add_description(
+        self,
+        worksheet: Worksheet,
+        description: str,
+        analytic_task_report: BaseAnalyticReportSettings,
+    ) -> None:
+        """Заполняет описание отчета."""
+        self.__add_row(worksheet, analytic_task_report, (description,))
+
     def __add_header(self, worksheet: Worksheet, analytic_task_report: BaseAnalyticReportSettings) -> None:
         """Заполняет первые строки в листе."""
         self.__add_row(worksheet, analytic_task_report, analytic_task_report.header_data)
@@ -80,12 +91,12 @@ class AnalyticReportBuilder:
         """Задаёт форматирование отчёта."""
         # задаём стиль для хэдера
         rows = list(worksheet.iter_rows())
-        header_cells = rows[0]
+        header_cells = rows[1]
         for cell in header_cells:
             cell.font = self.Styles.FONT_BOLD.value
             cell.alignment = self.Styles.ALIGNMENT_HEADER.value
         # задаём стиль для ячеек с данными
-        data_rows = rows[1:-1]
+        data_rows = rows[2:-1]
         for row in data_rows:
             for cell in row:
                 cell.font = self.Styles.FONT_STANDART.value
@@ -99,7 +110,7 @@ class AnalyticReportBuilder:
         for row in rows:
             for cell in row:
                 cell.border = self.Styles.BORDER.value
-        worksheet.column_dimensions["A"].width = self.Styles.WIDTH.value
+        worksheet.column_dimensions["B"].width = self.Styles.WIDTH.value
 
     class Styles(enum.Enum):
         FONT_BOLD = Font(name='Times New Roman', size=11, bold=True)
